requirements.txt
fastapi
uvicorn
streamlit
openai
opencv-python
numpy
pillow
mediapipe
pyttsx3
speechrecognition
requests
scikit-learn
pandas
matplotlib
seaborn
plotly
tensorflow
torch
transformers
flask
django
